refactor: move PageInfo to shared components, add wide variant
