+++ src/components/pages/doc-welcome/index.js
@@ -1,7 +1,7 @@
 export * from './cloud';
 export * from './features';
 export * from './help';
-export * from './page-info';
+export * from '../../shared/page-info';
 export * from './quickstart';
 export * from './showcase';
 export * from './what-is';
