--- src/components/pages/doc-welcome/page-info/index.js
@@ -1,1 +1,0 @@
-export { PageInfo } from './doc-welcome-page-info.view';
