+++ src/components/shared/page-info/index.js
@@ -1,0 +1,1 @@
+export { PageInfo } from './page-info.view';
