+++ src/templates/docs/javascript-api.js
@@ -3,10 +3,10 @@ import { HtmlContent } from 'components/blocks/html-content';
 import htmlStyles from 'components/blocks/html-content/html-content.module.scss';
 import jsApiStyles from 'components/pages/doc-javascript-api/doc-javascript-api.module.scss';
 import TableOfContents from 'components/pages/doc-page/table-of-contents';
-import { PageInfo } from 'components/pages/doc-welcome/page-info';
 import Blockquote from 'components/shared/blockquote';
 import { CodeGroup, Code, CodeInline } from 'components/shared/code';
 import CustomContentContainer from 'components/shared/custom-content-container';
+import { PageInfo } from 'components/shared/page-info';
 import TableWrapper from 'components/shared/table-wrapper';
 import docPageContent from 'components/templates/doc-page/doc-page-content/doc-page-content.module.scss';
 import LocaleProvider from 'contexts/locale-provider';
